git: make remote parsing a bit more robust
Sometimes the remote.<branch>.merge field is not set up to be fully
qualified.  It's not entirely clear when/why this happens, but since
repo itself handles this scenario, improve the pre-upload hook so it
also works.  We assume refs/heads/ is the implicit namespace as this
is what repo uses and is generally what everyone using git uses.

BUG=b:27812802
TEST=`repo upload` works w/merge=master & w/merge=refs/heads/master

Change-Id: I3e1fc728a61c8061791ce8f4138b8d565fb2c60f