README: rework config file headers
Now that there are multiple config file locations, putting the syntax
details under one of them doesn't make sense, so reparent to the main
config files section.

Also drop the code style on the header since gitiles can't handle it.
None of the sections will be rendered in the table of contents.

Bug: None
Test: Checked the new rendered output
Change-Id: Ie700989aa8d618a2ed2c011be374ef981c802961