Add an option to ignore merged commits
When uploading merge-commits, git log will list commits that are not
necessarily uploaded (since they have usually been already committed to
another branch). In order to avoid developers skipping verification on
these cases, this change adds an option to the config file that will add
--first-parent to the git log invocation, so pre-upload checks are only
performed on the commits that the developer intended.

Bug: 31926893
Test: repo upload
Change-Id: I76959906ed05b5cd09c4263fb54997b55dd6cfc3