hooks: Improve the clang-format wrapper script
This change leverages upstream changes in git-clang-format to restrict
diffs to a single commit instead of diffing from a commit to whatever
is in the working tree. That mode of operation does not lend itself to
an easy invocation to automatically fix any violations (since it
requires the use of the --diff flag), so a --fix flag was also added to
the wrapper script in order to automatically apply any changes that are
needed to comply.

Bug: 31305183
Test: repo upload
Change-Id: I835f2e1923803ff8ac9aee451078e6c1c1dcc070