Add support for automatically applying fixes
This change adds support for hooks to provide a mechanism to
automatically fix any errors found in the hook's execution. This will
currently only be attempted if the commit in question is the topmost in
the list of commits to be sent (i.e. if it is HEAD) and there is a
single fix available (since multiple fixes might interact badly with
each other).

This change also adds automatic fix support for the clang-format hook.

Bug: 32570902
Test: clang-format can now auto-fix formatting
Change-Id: I9214b7eb88548481feb639d8f8e70b3f57913dc0