Add a repo hook to validate TEST_MAPPING file.
TEST_MAPPING is a json file can exist anywhere in the Android Source
tree. It has a specific format to list tests to run in presubmit or
postsubmit (go/test-mapping). Instead of waiting for presubmit to fail
for an invalid TEST_MAPPING file, a repo hook is way more efficient
on validating the format.

Bug: 77805721
Test: unittest
Change-Id: Ia1440ce46029d7a59a915d22ccbb585bdf98298e