pre-upload: bubble up internal errors instead of exiting
Currently we exit if we hit an internal error (like invalid config file)
in a specific project repo.  Lets return the error to the caller so it
can continue running across all repos, and so it can summarize the run
in general with a standard error message.  Otherwise we don't currently
include a whole lot of info in the output which is confusing.

Bug: 124462370
Test: running `repo upload` across multiple repos doesn't exit immediately
Change-Id: I36022286750958f49a761d547b341447c520b411