pre-upload: require Python 2.7.5+ and 3.4+
We def don't work with Python 2.6 or earlier, and versions before
2.7.5 have a signal bug that we've been working around.  Since that
was released in 2013 (6 years ago), lets just require newer Python
so we can drop the workarounds.

Similarly, we def don't work with Python 3.3 or earlier, and versions
before 3.4 have that same signaling bug.  It's unclear if we even run
on Python 3.4, but lets set the bar low for now.

Bug: 142671918
Test: `repo upload` passes
Change-Id: Ie259d4f67f55a18216927a911c04e9e00e0562ef