utils: sudo_run: delete
We were using this function in only one error case in one scenario
that shouldn't really happen.  Expecting to get access to root in
general & invoking sudo is weird for tests, so don't even try to
support it.  If we hit this situation, let's abort and make the
user clean things up.

Bug: None
Test: preupload unittests pass
Change-Id: Ie1ac33b477b46ddacf372b661f160421ec61049b