Fix regex in commit_msg_relnote_for_current_txt repo hook
The current (incorrect) regex only enforces a Relnote
tag for current.txt files in the root directory of the
repo.  However, current.txt can be in any repo subdirectory,
so we need to enforce it everywhere.

This change removes the up carrot in the regex that
enforces matches must start at the line start.

Bug: 159449967
Test: ./pre-upload.py (with commit_msg_relnote_for_current_txt = true)
Change-Id: I7555398df6edaf410af60ce0def05894ef53ea11