hooks: rustfmt: handle multiple commits
The rustfmt hook was relying on ${PREUPLOAD_FILES} which is only correct
when evaluating HEAD. Update the hook to parse the correct version of
the file.

Before 2.0, rustfmt does not support --check when passing the source via
stdin. Handle this case by comparing the input with the output and
report a command to fix the issue to the user.

Bug: 164111102
Test: Added multiples changes to system/security (typo and incorrect
  indentation). Run pre-upload.py manually.
Change-Id: Ibd1bfcb9e630f0b21abfb1e1feb1248fcddf2236