Add [Builtin Hooks Exclude Paths] section
There is currently no convenient option to enable a hook globally if
some projects will fail the processing. The recommended setup is to
enable the hook within each project's repository (using PREUPLOAD.cfg).
This creates inconsistencies for large codebase. Adds a new
configuration section to explicitly exclude some projects during the
processing of a hook.

The intent of this change is to enable rustfmt globally in AOSP,
except for some paths (e.g. external/, vendor/).

Test: Modified GLOBAL-PREUPLOAD.cfg to enable the new option,
  manually creates changes and review output of pre-upload.py
Bug: 160223496
Change-Id: I94dbbf0ce2e6b58c4d4b4fc89c56a2a87543d878